Fix type mismatch errors in min/max calls
- Add explicit uint16_t casts in LedEffects.h min() call
- Replace max() calls with conditional expressions for brightness calculations
- Fix type mismatch in Config.h calculateSpeed() function
- Ensures compatibility with strict C++ template type checking

+++ lib/PongCommon/Config.h
@@ -188,7 +188,7 @@ inline uint16_t calculateSpeed(const GameConfig& config, uint16_t round) {
     }
 
     // Apply speed cap
-    return max(config.maxSpeed, speed);
+    return max((uint16_t)config.maxSpeed, (uint16_t)speed);
 }
 
 // Calculate average reaction time
